ml_pipeline_project/airflow_setup/dags/ml_pipeline/requirements.txt
pandas
scikit-learn
# Airflow環境で実行する場合に必要
apache-airflow
joblib